Extract type for Matchers implements
